demo/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
idna==2.10
markup==0.2
publicip==1.0.1
PyJWT==1.7.1
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
gunicorn==20.0.4
psycopg2==2.8.6
